backend/requirements.txt

# Core providers used by the service adapters
openai>=0.27.0
httpx>=0.24.0
anthropic>=0.8.0
google-cloud-aiplatform>=1.30.0

# Optional: load env vars from .env during development
python-dotenv>=1.0.0

fastapi
uvicorn[standard]
pydantic
sqlalchemy
psycopg2-binary
alembic
httpx
python-dotenv
celery[redis]
redis
python-multipart
requests